+++ src/memex_logging/ws/resource/analytic.py
@@ -37,13 +37,13 @@ class AnalyticsResourceBuilder(object):
     @staticmethod
     def routes(es: Elasticsearch):
         return [
-            (AnalyticsPerformer, '/analytic', (es,)),
+            (AnalyticInterface, '/analytic', (es,)),
             (GetNoClickPerUser, '/analytic/usercount', (es,)),
             (GetNoClickPerEvent, '/analytic/eventcount', (es,))
         ]
 
 
-class AnalyticsPerformer(Resource):
+class AnalyticInterface(Resource):
 
     def __init__(self, es: Elasticsearch):
         self._es = es
@@ -60,7 +60,14 @@ class AnalyticsPerformer(Resource):
 
         project = request.args.get('project', None)
         index_name = Utils.generate_index("analytic", project=project)
-        response = self._es.search(index=index_name, body={"query": {"match": {"staticId.keyword": static_id}}})
+        try:
+            response = self._es.search(index=index_name, body={"query": {"match": {"staticId.keyword": static_id}}})
+        except Exception as e:
+            logger.exception(f"Analytic with static_id [{static_id}] failed to be retrieved", exc_info=e)
+            return {
+                "status": "Internal server error: could not retrieved the analytic",
+                "code": 500
+            }, 500
 
         if response['hits']['total']['value'] == 0:
             logger.debug("Resource not found")
@@ -116,6 +123,32 @@ class AnalyticsPerformer(Resource):
 
         update_analytic.delay(static_id=static_id)
         return {"staticId": static_id}, 200
+
+    def delete(self):
+        static_id = request.args.get('staticId')
+        logger.info(f"Deleting analytic with static_id: {static_id}")
+        if static_id == "" or static_id is None:
+            logger.debug("Missing required staticId parameter")
+            return {
+                "status": "Malformed request: missing required parameter `staticId`",
+                "code": 400
+            }, 400
+
+        project = request.args.get('project', None)
+        index_name = Utils.generate_index("analytic", project=project)
+        try:
+            self._es.delete_by_query(index=index_name, body={"query": {"match": {"staticId.keyword": static_id}}})
+        except Exception as e:
+            logger.exception(f"Analytic with static_id [{static_id}] failed to be deleted", exc_info=e)
+            return {
+                "status": "Internal server error: could not delete the analytic",
+                "code": 500
+            }, 500
+
+        return {
+            "status": "Ok: analytic deleted",
+            "code": 200
+        }, 200
 
 
 class GetNoClickPerUser(Resource):